done for you connect stuff

+++ database/migrations/2021_03_01_141343_create_worker_comments_table.php
@@ -17,10 +17,10 @@ class CreateWorkerCommentsTable extends Migration
             $table->id();
             $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            
+
             $table->bigInteger('parent_id')->unsigned();
             $table->text('body');
-            $table->integer('commentable_id')->unsigned();
+            $table->bigInteger('commentable_id')->unsigned();
             $table->string('commentable_type');
             $table->timestamps();
         });
